requirements.txt
# Core dependencies
anthropic
claude-agent-sdk
mcp
requests
python-dotenv

# Web backend
fastapi
uvicorn[standard]

# Deploy
paramiko